Fix #21
Include also the events (and not just the event instances) in the set
of relevant instances.

The issue is that some event can not have instances (e.g. there is no
adequate mapping in the spec).
This means that the event, when executed, will not change the
permitted/prohibited status of another event/callin (and its status
would also not be changed).

However, the event can still be executed by the scheduler, and hence
it can executes some callins.
Thus, the event cannot be safely removed from the trace.