Consider framework inheritance in the bindings
A callback message can be matched by several specifications, and the
method call in the specification can be a framework class or an
interface implemented by the callback message.

The substitution did not replace the framework class name or the
interface name in the instantiated specification. This way it was not
possible to find the right variable in the encoding (the variables are
linked to the concrete messages).

The commit fix this.