Implement pruning of assignments for the grounding
- Fix the generation of reasons (subsets of a complete assignments to
the free variables) that is sufficient to ground a specification.
- When the grounded specification is false, we use the reason to prune
the remaining set of assignments.
The idea is that with that reason (subset of assignments) we cannot
ground a specification. Thus, we can remove it from the remaning
assignments.
The effect is to prune sets of assignments at once from the
enumeration.